# UserService/api-docs.json
{"openapi":"3.0.1","info":{"title":"Demo API","description":"Spring Boot RestFul AlbumSystem API Demo by Reene","termsOfService":"https://github.com/Reene444","contact":{"name":"album-system-api","url":"https://github.com/Reene444","email":"reene44444@gmail.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"Verions 1.0"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"tags":[{"name":"Albumm Controller","description":"Controller for album and photo management"},{"name":"Test","description":"The Test API."},{"name":"Auth Controller","description":"Controller for Account management"}],"paths":{"/api/v1/auth/users/{user_id}/update-authorities":{"put":{"tags":["Auth Controller"],"summary":"Update authorities","operationId":"update_auth","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthoritiesDTO"}}},"required":true},"responses":{"200":{"description":"Update authorities","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"400":{"description":"Invalid user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/auth/profile/update-password":{"put":{"tags":["Auth Controller"],"summary":"Update profile","operationId":"update_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDTO"}}},"required":true},"responses":{"200":{"description":"Profile of users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/update":{"put":{"tags":["Albumm Controller"],"summary":"Update an Album","operationId":"update_Album","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"400":{"description":"Please add a valid description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"204":{"description":"Album update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/update":{"put":{"tags":["Albumm Controller"],"summary":"Update a photo","operationId":"update_Photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoPayloadDTO"}}},"required":true},"responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"400":{"description":"Please add a valid description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"204":{"description":"Album update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/auth/users/add":{"post":{"tags":["Auth Controller"],"summary":"Add a new user","operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDTO"}}},"required":true},"responses":{"400":{"description":"Please enter a vaild email and password lenght between 6 to 20 characters","content":{"application/josn":{"schema":{"type":"string"}}}},"200":{"description":"Account added","content":{"application/josn":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"application/josn":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["Auth Controller"],"operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/albums/{album_id}/upload-photos":{"post":{"tags":["Albumm Controller"],"summary":"Upload photo into album","operationId":"photos","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["files"],"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"400":{"description":"Please check the payload or token","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object"}}}}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/add":{"post":{"tags":["Albumm Controller"],"summary":"Add an Album api","operationId":"addAlbum","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"201":{"description":"Account added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"400":{"description":"Please add valid name a description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/test":{"get":{"tags":["Test"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"rest-test1-api":[]}]}},"/api/v1/auth/users":{"get":{"tags":["Auth Controller"],"summary":"List user api","operationId":"Users","responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/auth/profile":{"get":{"tags":["Auth Controller"],"summary":"View profile","operationId":"profile","responses":{"200":{"description":"Profile of users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums":{"get":{"tags":["Albumm Controller"],"summary":"List of album api","operationId":"albums","responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"200":{"description":" List of albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}":{"get":{"tags":["Albumm Controller"],"summary":"List album by album ID","operationId":"albums_by_id","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"200":{"description":" List of albums","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-thumbnail":{"get":{"tags":["Albumm Controller"],"operationId":"downloadThumbnail","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-photo":{"get":{"tags":["Albumm Controller"],"operationId":"downloadPhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"album-system-api":[]}]}},"/":{"get":{"tags":["home-controller"],"operationId":"demo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/profile/delete":{"delete":{"tags":["Auth Controller"],"summary":"Delete profile","operationId":"delete_profile","responses":{"200":{"description":"Profile of users","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Token error","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Token missing","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/delete":{"delete":{"tags":["Albumm Controller"],"summary":"delete a photo","operationId":"delete_photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Photo delete","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"album-system-api":[]}]}},"/api/v1/albums/{album_id}/delete":{"delete":{"tags":["Albumm Controller"],"summary":"delete a photo","operationId":"delete_album","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Album deleted","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"studyeasy-demo-api":[]}]}}},"components":{"schemas":{"AuthoritiesDTO":{"required":["authorities"],"type":"object","properties":{"authorities":{"type":"string","description":"Authorities","example":"USER"}}},"AccountViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}},"PasswordDTO":{"required":["password"],"type":"object","properties":{"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"password"}}},"AlbumPayloadDTO":{"required":["description","name"],"type":"object","properties":{"name":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"}}},"AlbumViewDTO":{"required":["description","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"},"photos":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDTO"}}}},"PhotoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"fileName":{"type":"string"},"download_link":{"type":"string"}}},"PhotoPayloadDTO":{"required":["description","name"],"type":"object","properties":{"name":{"type":"string","description":"Photo name","example":"Travel"},"description":{"type":"string","description":"Description of the photo","example":"Description"}}},"PhotoViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"AccountDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"admin@admin.com"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"password"}}},"UserLoginDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"user@user.com"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"pass987"}}},"TokenDTO":{"type":"object","properties":{"token":{"type":"string"}}},"ProfileDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}}},"securitySchemes":{"album-system-api":{"type":"http","in":"header","scheme":"bearer"}}}}